spy: fix parsed VAA type

+++ solana/ts/auction-participant/vaaAuctionRelayer/app.ts
@@ -1,12 +1,12 @@
 import { Connection, Keypair, PublicKey } from "@solana/web3.js";
 import "dotenv/config";
 import * as fs from "fs";
-import { MatchingEngineProgram } from "../../src/matchingEngine";
-import { PreparedTransaction } from "../../src";
-import * as utils from "../utils";
 import * as winston from "winston";
+import { PreparedTransaction } from "../../src";
+import { MatchingEngineProgram } from "../../src/matchingEngine";
 import { VaaSpy } from "../../src/wormhole/spy";
 import { CachedBlockhash } from "../containers";
+import * as utils from "../utils";
 
 const MATCHING_ENGINE_PROGRAM_ID = "mPydpGUWxzERTNpyvTKdvS7v8kvw5sgwfiP8WQFrXVS";
 const USDC_MINT = new PublicKey("4zMMC9srt5Ri5X14GAgXhaHii3GnPAEERYPJgZJDncDU");
@@ -81,9 +81,7 @@ async function main(argv: string[]) {
 
             // Start a new auction if this is a fast VAA.
             logicLogger.debug(`Attempting to parse FastMarketOrder, sequence=${parsed.sequence}`);
-            const fastOrder = utils.tryParseFastMarketOrder(
-                Buffer.from(parsed.payload as Uint8Array),
-            );
+            const fastOrder = utils.tryParseFastMarketOrder(Buffer.from(parsed.payload));
             if (fastOrder !== undefined) {
                 const unprocessedTxns = await utils.handlePlaceInitialOffer(
                     connection,
@@ -102,9 +100,7 @@ async function main(argv: string[]) {
             }
         } else {
             logicLogger.debug(`Attempting to parse SlowOrderResponse, sequence=${parsed.sequence}`);
-            const slowOrderResponse = utils.tryParseSlowOrderResponse(
-                Buffer.from(parsed.payload as Uint8Array),
-            );
+            const slowOrderResponse = utils.tryParseSlowOrderResponse(Buffer.from(parsed.payload));
             if (slowOrderResponse !== undefined) {
                 const unprocessedTxns = await utils.handleSettleAuction(
                     connection,
